requirements.txt
pytorch==1.4.0
torchvision==0.5.0
numpy==1.17.3
matplotlib==3.1.2